third_party/blink/web_tests/html/selectmenu/selectmenu-slot-warning-button-expected.txt

CONSOLE WARNING: A <selectmenu>'s default button was removed and a new one was not provided. This <selectmenu> will not be fully functional.
This is a testharness.js-based test.
PASS selectmenu-slot-warning-button
Harness: the test ran to completion.
